fix for BigFloat
When s0 is a BigFloat, the expression ((k+3)/(k+4))*s0 converts ((k+3)/(k+4)) to a binary64, then when it is multiplied by s0, it is promoted to a BigFloat. That's not what we want.

I could do either  ((k+3)//(k+4))*s0 or ((k+3)*s0)/(k+4). Let's try ((k+3)*s0)/(k+4).