initial effort at fixing the new code
This is an initial effort at fixing the new code. The code was using the wrong value of alpha.  I've extended the Maxima worksheet to derive these results.  The Maxima worksheet needs some work still.

There is still a great deal of work to be done.  Although, the series used by the new code converges off [1,infinity), it can converge slowly.  So we still need to  do the transformations to get sufficiently fast convergence.