experimental code mostly works
Work on finishing the experimental new function.   I've marked a few tests as broken--these are tests that pass with the old, but not the experimental code.

The experimental code is slower--I need a theoretical way to know which method is best. But I'd guess that the experimental code could be sped up somewhat (eliminate redundant arithmetic).